+++ vite.config.js
@@ -16,6 +16,7 @@ export default defineConfig({
             '@utils': fileURLToPath(new URL('./src/utils', import.meta.url)),
             '@assets': fileURLToPath(new URL('./src/assets', import.meta.url)),
             '@services': fileURLToPath(new URL('./src/services', import.meta.url)),
+            '@hooks': fileURLToPath(new URL('./src/hooks', import.meta.url)),
         },
     },
     server: {